Chrysalis.device: Start moving hardware knowledge out
In the end, Chrysalis.device should be hardware-agnostics, and rely on other
parts of the code for supplying the necessary bits. For this reason, the
`detect()` function was changed so that the logic of deciding whether a device
is supported or not, lives outside of the module.

As a consequence, the promise now emits multiple "device-detected" events, one
for each device, instead of one event for all devices combined.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>