command: Much improved device I/O
Instead of waiting for a set timeout for a full response after sending a
command, use an on-data handler. In practice this means we collect all the data,
and process it asynchronously, by using a command queue.

The end result is a lot faster IO, more responsive UI updates, and a much less
fragile IO layer.

Fixes #58.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>